Create the 1st recipe to act as the boilerplate
Insert the initial code and add link to the index page; a heading, followed by recipe image, description title, then a description, followed by ingredients title and an unordered list of the ingredients. Then finally, a steps title followed by an unordered list of the steps to make the recipe.